Introduce a "transactional" diagnostic consumer.
Swift compiler and ClangImporter and LLDB operations can produce
diagnostics, which are stored in StoringDiagnosticConsumer. Operations
such as expression evaluation of module imports may produce multiple
non-fatal error diagnostics.  Prior to this patch, the expression
parser would clear the diagnostic consumer before running. This had
the unwanted side-effect of sometimes clearing out unreported earlier
diagnostics.

This patch introduces SwiftASTContext::forkDiagnosticConsumer() which
returns an RAII object that behaves like a diagnostic consumer, can
tell whether diagnostics where produced during its lifetime, and
resets the StoringDiagnosticConsumer to its previous state upon
destruction. This way

- unreported diagnostics don't get cleared by the expression parser
- the expression parser has precise access to only the diagnostics it produced
- while still being able to report persistent errors
  (such as clang import failures) that happened before the expression
  parser started.

I'm not happy by the amount of code needed to make this work, or the
specific implementation, but at least the interface provided to the
outside is nice and clean.
